feat(auth): more work on auth
