+++ src/components/Router/index.jsx
@@ -5,14 +5,14 @@ import UserDashboard from '../../pages/UserDashboard';
 import DashboardLayout from '../Layout/DashboardLayout';
 
 const AppRouter = () => {
-    return (
-        <BrowserRouter>
-          <Switch>
-            <Route exact path="/" component={Home} />
-            <DashboardLayout path="/dashboard" component={UserDashboard} />
-          </Switch>
-        </BrowserRouter>
-    )
-}
+  return (
+    <BrowserRouter>
+      <Switch>
+        <Route exact path="/" component={Home} />
+        <DashboardLayout path="/dashboard" component={UserDashboard} />
+      </Switch>
+    </BrowserRouter>
+  );
+};
 
 export default AppRouter;
