+++ src/pages/UserDashboard.jsx
@@ -1,8 +1,11 @@
+/* eslint-disable no-shadow */
 /* eslint-disable react/prop-types */
 import React, { useEffect } from 'react';
+import axios from 'axios';
 import { connect } from 'react-redux';
 import styled from 'styled-components';
-import { actions } from '../state/actions';
+import { Spin } from 'antd';
+import { loginUser } from '../state/actions/auth';
 import ReviewList from '../components/ReviewList';
 
 const StyledH1 = styled.h1`
@@ -10,8 +13,31 @@ const StyledH1 = styled.h1`
   padding-left: 9px;
 `;
 
-const UserDashboard = props => {
-  return (
+const UserDashboard = ({ authState: { isLoggingIn }, loginUser }) => {
+  useEffect(async () => {
+    const urlParams = new URLSearchParams(window.location.search);
+    const code = urlParams.get('code');
+
+    const getUserDetails = async () => {
+      const {
+        data: {
+          user_id: userId,
+          user: { name, email, image_1024: profilePicture },
+        },
+      } = await axios.get(
+        `https://slack.com/api/oauth.access?client_id=${process.env.REACT_APP_CLIENT_ID}&client_secret=${process.env.REACT_APP_CLIENT_SECRET}&code=${code}&redirect_uri=${process.env.REACT_APP_REDIRECT_URI}`
+      );
+      window.history.replaceState(null, null, window.location.pathname);
+      await loginUser(userId, name, email, profilePicture);
+    };
+
+    if (code) {
+      await getUserDetails();
+    }
+  }, []);
+  return isLoggingIn ? (
+    <Spin />
+  ) : (
     <div>
       <StyledH1>Latest Reviews</StyledH1>
       <ReviewList />
@@ -19,4 +45,4 @@ const UserDashboard = props => {
   );
 };
 
-export default connect(state => state, actions)(UserDashboard);
+export default connect(state => state, { loginUser })(UserDashboard);
